kernel: do not panic on failed initialization
When family or netdev notifier registration fails, do not panic, just
WARN() so that a warning is logged and kernel tainted but the system can
boot. If family registration fails, set a flag so that ethtool_notify()
bails out immediately. (This is the only external entry point which can
be called without family registered.)

Rebase to net-next commit 68cc2999f692.