kernel: reorder patches to prepare for v6
None of the patches to be submitted in v6 of first part uses the
nla_put_bitfield32() helper so move the patch adding it just before the
first patch using it. In the end, we may use arbitrary size bitmaps even
for the attributes where NLA_BITFIELD32 is used now so that the helper
wouldn't be used at all.