ethtool: replace the outdated series
Like kernel, the ethtool series is seriously lacking behind the state of
the work and also had to be reduced to only the patches targeted for
next submission.

Replace the old series by current work in progress. This series
implements "ethtool <dev>" and "ethtool -s <dev> ..." using the netlink
API implemented in kernel 5.6-rc1.

The series is a result of a fresh refactoring and reorganization and is,
in this form, almost untested at the moment.

Rebase to ethtool commit 7dc0af7e386e.